Move PU-specific test entities into their own packages
(cherry picked from commit 9d58018bfc676db6cfe3b33a6c1dcef071a19f86)
