+++ integration-tests/jpa-postgresql-withxml/src/main/java/io/quarkus/it/jpa/postgresql/defaultpu/SequencedAddress.java
@@ -1,4 +1,4 @@
-package io.quarkus.it.jpa.postgresql;
+package io.quarkus.it.jpa.postgresql.defaultpu;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
