+++ integration-tests/jpa-postgresql-withxml/src/main/resources/application.properties
@@ -3,7 +3,7 @@ quarkus.datasource.password=hibernate_orm_test
 quarkus.datasource.jdbc.url=${postgres.url}
 quarkus.datasource.jdbc.max-size=8
 
-quarkus.hibernate-orm.packages=io.quarkus.it.jpa.postgresql
+quarkus.hibernate-orm.packages=io.quarkus.it.jpa.postgresql.defaultpu
 quarkus.hibernate-orm.database.generation=drop-and-create
 
 #Necessary for assertions in JPAFunctionalityInGraalITCase:
@@ -15,3 +15,4 @@ quarkus.native.additional-build-args=-J-Dio.quarkus.jdbc.postgresql.graalvm.diag
 # Define non-default PU so that we can configure a custom XML format mapper. The default PU is using the default mapper.
 quarkus.hibernate-orm."other".datasource=<default>
 quarkus.hibernate-orm."other".packages=io.quarkus.it.jpa.postgresql.otherpu
+quarkus.hibernate-orm."other".database.generation=drop-and-create
