+++ integration-tests/jpa-postgresql/src/main/resources/application.properties
@@ -3,12 +3,14 @@ quarkus.datasource.password=hibernate_orm_test
 quarkus.datasource.jdbc.url=${postgres.url}
 quarkus.datasource.jdbc.max-size=8
 
-quarkus.hibernate-orm.packages=io.quarkus.it.jpa.postgresql
+quarkus.hibernate-orm.packages=io.quarkus.it.jpa.postgresql.defaultpu
 quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.database.generation.create-schemas=true
 # Define non-default PU so that we can configure a custom JSON format mapper. The default PU is using the default mapper.
 quarkus.hibernate-orm."other".datasource=<default>
 quarkus.hibernate-orm."other".packages=io.quarkus.it.jpa.postgresql.otherpu
+quarkus.hibernate-orm."other".database.generation=drop-and-create
+quarkus.hibernate-orm."other".database.generation.create-schemas=true
 
 #Necessary for assertions in JPAFunctionalityInGraalITCase:
 quarkus.native.enable-reports=true
